+++ app/Http/Controllers/PartNumberController.php
@@ -12,9 +12,9 @@ class PartNumberController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        PartNumber::all();
+        return PartNumber::all();
     }
 
     /**
@@ -25,8 +25,34 @@ class PartNumberController extends Controller
      */
     public function store(Request $request)
     {
+        if (!is_array($request->all())) {
+            return ['error' => 'request must be an array'];
+        }
+        
+        //Reglas de Validacion
+        $rules = [
+            'name'    => 'required',
+            'description' => 'required'
+        ];
+        
+        try {
+        //Se ejecuta el validador
+        $validator = \Validator::make($request->all(), $rules);
+        if($validator->fails()){
+            return [
+                'created' => false,
+                'errors' => $validator->errors()->all()
+                ];
+            
+        }
+        
         PartNumber::create($request->all());
         return ['created' => true];
+        
+        }catch (Exception $e) {
+            \Log::info('Error creating PartNumber: '.$e);
+            return \Response::json(['created' => false], 500);
+        }
     }
 
     /**
@@ -37,7 +63,7 @@ class PartNumberController extends Controller
      */
     public function show($id)
     {
-         return PartNumber::findOrFail($id);
+       return PartNumber::findOrFail($id);
     }
 
 
@@ -50,9 +76,35 @@ class PartNumberController extends Controller
      */
     public function update(Request $request, $id)
     {
+        
+        if (!is_array($request->all())) {
+            return ['error' => 'request must be an array'];
+        }
+        
+        //Reglas de Validacion
+        $rules = [
+            'name'  => 'required',
+            'description' => 'required'
+        ];
+        
+         try {
+        //Se ejecuta el validador
+        $validator = \Validator::make($request->all(), $rules);
+        if($validator->fails()){
+            return [
+                'updated' => false,
+                'errors' => $validator->errors()->all()
+            ];
+        }
+        
         $partnumber = PartNumber::findOrFail($id);
         $partnumber->update($request->all());
         return ['updated' => true];
+             
+            }catch (Exception $e) {
+            \Log::info('Error updating the  PartNumber: '.$e);
+            return \Response::json(['updated' => false], 500);
+        }
     }
     
 
@@ -62,10 +114,31 @@ class PartNumberController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+        //Reglas de Validacion
+        $rules = [
+            'id'  => 'required|exist:partnumbers,id'
+        ];
+        
+        try {
+        //Se ejecuta el validador
+        $validator = \Validator::make($request->all(), $rules);
+        if($validator->fails()){
+            return [
+                'deleted' => false,
+                'errors' => $validator->errors()->all()
+            ];
+        }
+        
+        
         PartNumber::destroy($id);
         return ['deleted' => true];
+            
+            }catch (Exception $e) {
+            \Log::info('Error deleting the selected PartNumber: '.$e);
+            return \Response::json(['deleted' => false], 500);
+        }
     }
     /**
      * Display the partnumber with the name.
@@ -73,8 +146,9 @@ class PartNumberController extends Controller
      * @param  string  $name
      * @return \Illuminate\Http\Response
      */
-    public function getByName($name){
-
+    public function getByName( $name){
+        
          return PartNumber::where('name','=', $name)->get();
-    }
+            
+           }
 }
