+++ app/Http/Controllers/RiskReleaseController.php
@@ -14,7 +14,7 @@ class RiskReleaseController extends Controller
      */
     public function index()
     {
-        return RiskRelease::all();
+       return RiskRelease::all();
     }
 
     /**
@@ -25,8 +25,36 @@ class RiskReleaseController extends Controller
      */
     public function store(Request $request)
     {
-        RiskRelease::create($request->all());
-        return ['created' => true];
+        if (!is_array($request->all())) {
+            return ['error' => 'request must be an array'];
+        }
+
+         // Creamos las reglas de validación
+        $rules = [
+            'description'      => 'required',
+            'user_id' => 'required|exist:users,id',
+            'init_date' => 'required',
+            'end_date' => 'required',
+            'created_at' => 'required',
+            'updated_at' => 'required'
+            ];
+
+       try {
+            // Ejecutamos el validador y en caso de que falle devolvemos la respuesta
+            $validator = \Validator::make($request->all(), $rules);
+            if ($validator->fails()) {
+                return [
+                    'created' => false,
+                    'errors'  => $validator->errors()->all()
+                ];
+            }
+ 
+            RiskRelease::create($request->all());
+            return ['created' => true];
+        } catch (Exception $e) {
+            \Log::info('Error creating risk release: '.$e);
+            return \Response::json(['created' => false], 500);
+        }
     }
 
     /**
@@ -37,7 +65,7 @@ class RiskReleaseController extends Controller
      */
     public function show($id)
     {
-        return RiskRelease::findOrFail($id);
+     return RiskRelease::findOrFail($id);
     }
 
     /**
@@ -49,9 +77,38 @@ class RiskReleaseController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (!is_array($request->all())) {
+            return ['error' => 'request must be an array'];
+        }
+        
+                //Reglas de Validacion
+        $rules = [
+            'description'      => 'required',
+            'user_id' => 'required|exist:users,id',
+            'init_date' => 'required',
+            'end_date' => 'required',
+            'created_at' => 'required',
+            'updated_at' => 'required'
+        ];
+        
+        try {
+        //Se ejecuta el validador
+        $validator = \Validator::make($request->all(), $rules);
+        if($validator->fails()){
+            return [
+                'updated' => false,
+                'errors' => $validator->errors()->all()
+            ];
+        }
+        
         $riskrelease = RiskRelease::findOrFail($id);
         $riskrelease->update($request->all());
         return ['updated' => true];
+            
+            }catch (Exception $e) {
+            \Log::info('Error updating the requested risk release: '.$e);
+            return \Response::json(['updated' => false], 500);
+        }
     }
     
 
@@ -63,7 +120,27 @@ class RiskReleaseController extends Controller
      */
     public function destroy($id)
     {
+        //Reglas de Validacion
+        $rules = [
+            'id'  => 'required|exist:riskreleases,id'
+        ];
+        
+        try {
+        //Se ejecuta el validador
+        $validator = \Validator::make($request->all(), $rules);
+        if($validator->fails()){
+            return [
+                'deleted' => false,
+                'errors' => $validator->errors()->all()
+            ];
+        }
+        
         RiskRelease::destroy($id);
         return ['deleted' => true];
+            
+            }catch (Exception $e) {
+            \Log::info('Error deleting the selected risk release: '.$e);
+            return \Response::json(['deleted' => false], 500);
+        }
     }
 }
